moved API into separate file and cleaned up
