+++ src/components/Main.css
@@ -31,3 +31,7 @@
 h1 {
     text-align: center;
 }
+
+select{
+    width:250px;
+}
